refactor rating crud [regroup]
