+++ handler/rating/delete.js
