+++ handler/rating/getRatingByContent.js
